style cleanup

+++ IPTables.Net/Iptables/IpSet/IpSetSet.cs
@@ -19,13 +19,13 @@ namespace IPTables.Net.Iptables.IpSet
     {
         #region Fields
 
-        private String _name;
+        private string _name;
         private IpSetType _type;
         private int _timeout;
         private string _family = "inet";
         private int _hashSize = 1024;
         private PortOrRange _bitmapRange = new PortOrRange(1, 65535, '-');
-        private UInt32 _maxElem = 65536;
+        private uint _maxElem = 65536;
         private HashSet<IpSetEntry> _entries;
         private IpTablesSystem _system;
         private IpSetSyncMode _syncMode = IpSetSyncMode.SetAndEntries;
@@ -34,21 +34,18 @@ namespace IPTables.Net.Iptables.IpSet
 
         internal string InternalName
         {
-            set { _name = value; }
+            set => _name = value;
         }
 
         #endregion
 
         #region Properties
 
-        public string Name
-        {
-            get { return _name; }
-        }
+        public string Name => _name;
 
         public IpSetType Type
         {
-            get { return _type; }
+            get => _type;
             set
             {
                 _type = value;
@@ -58,37 +55,34 @@ namespace IPTables.Net.Iptables.IpSet
 
         public int Timeout
         {
-            get { return _timeout; }
-            set { _timeout = value; }
+            get => _timeout;
+            set => _timeout = value;
         }
 
-        public UInt32 MaxElem
+        public uint MaxElem
         {
-            get { return _maxElem; }
-            set { _maxElem = value; }
+            get => _maxElem;
+            set => _maxElem = value;
         }
 
         public int HashSize
         {
-            get { return _hashSize; }
-            set { _hashSize = value; }
+            get => _hashSize;
+            set => _hashSize = value;
         }
 
-        public virtual HashSet<IpSetEntry> Entries
-        {
-            get { return _entries; }
-        }
+        public virtual HashSet<IpSetEntry> Entries => _entries;
 
         public IpSetSyncMode SyncMode
         {
-            get { return _syncMode; }
-            set { _syncMode = value; }
+            get => _syncMode;
+            set => _syncMode = value;
         }
 
         public string Family
         {
-            get { return _family; }
-            set { _family = value; }
+            get => _family;
+            set => _family = value;
         }
 
         public string[] TypeComponents
@@ -101,27 +95,21 @@ namespace IPTables.Net.Iptables.IpSet
             }
         }
 
-        public IpTablesSystem System
-        {
-            get { return _system; }
-        }
+        public IpTablesSystem System => _system;
 
-        public List<String> CreateOptions
-        {
-            get { return _createOptions; }
-        }
+        public List<string> CreateOptions => _createOptions;
 
         public PortOrRange BitmapRange
         {
-            get { return _bitmapRange; }
-            set { _bitmapRange = value; }
+            get => _bitmapRange;
+            set => _bitmapRange = value;
         }
 
         #endregion
 
         #region Constructor
 
-        public IpSetSet(IpSetType type, string name, int timeout, String family, IpTablesSystem system,
+        public IpSetSet(IpSetType type, string name, int timeout, string family, IpTablesSystem system,
             IpSetSyncMode syncMode, List<string> createOptions = null, HashSet<IpSetEntry> entries = null)
         {
             _type = type;
@@ -131,10 +119,12 @@ namespace IPTables.Net.Iptables.IpSet
             _system = system;
             _syncMode = syncMode;
             _createOptions = createOptions == null ? new List<string>() : createOptions.ToList();
-            _entries = entries == null ? new HashSet<IpSetEntry>(IpSetEntryKeyComparer.Instance) : entries.ToHashSet(IpSetEntryKeyComparer.Instance);
+            _entries = entries == null
+                ? new HashSet<IpSetEntry>(IpSetEntryKeyComparer.Instance)
+                : entries.ToHashSet(IpSetEntryKeyComparer.Instance);
         }
 
-        public IpSetSet(IpSetType type, string name, int timeout, String family, IpTablesSystem system,
+        public IpSetSet(IpSetType type, string name, int timeout, string family, IpTablesSystem system,
             IpSetSyncMode syncMode, PortOrRange bitmapRange, List<string> createOptions = null,
             HashSet<IpSetEntry> entries = null)
         {
@@ -145,7 +135,9 @@ namespace IPTables.Net.Iptables.IpSet
             _system = system;
             _syncMode = syncMode;
             _createOptions = createOptions == null ? new List<string>() : createOptions.ToList();
-            _entries = entries == null ? new HashSet<IpSetEntry>(IpSetEntryKeyComparer.Instance) : entries.ToHashSet(IpSetEntryKeyComparer.Instance);
+            _entries = entries == null
+                ? new HashSet<IpSetEntry>(IpSetEntryKeyComparer.Instance)
+                : entries.ToHashSet(IpSetEntryKeyComparer.Instance);
             _bitmapRange = bitmapRange;
         }
 
@@ -160,72 +152,53 @@ namespace IPTables.Net.Iptables.IpSet
 
         #region Methods
 
-        public String GetCommand()
+        public string GetCommand()
         {
-            String type = IpSetTypeHelper.TypeToString(_type);
-            String command = String.Format("{0} {1}", _name, type);
+            var type = IpSetTypeHelper.TypeToString(_type);
+            var command = string.Format("{0} {1}", _name, type);
 
             if ((_type & IpSetType.Hash) == IpSetType.Hash)
-            {
                 command += " family " + _family;
-            }
-            else if ((_type & IpSetType.Bitmap) == IpSetType.Bitmap)
-            {
-                command += " range " + _bitmapRange;
-            }
+            else if ((_type & IpSetType.Bitmap) == IpSetType.Bitmap) command += " range " + _bitmapRange;
 
             if ((_type & (IpSetType.Hash | IpSetType.CtHash)) != 0)
-            {
-                command += String.Format(" hashsize {0} maxelem {1}", _hashSize, _maxElem);
-            }
+                command += string.Format(" hashsize {0} maxelem {1}", _hashSize, _maxElem);
 
-            if (_timeout > 0)
-            {
-                command += " timeout " + _timeout;
-            }
+            if (_timeout > 0) command += " timeout " + _timeout;
 
-            foreach (var co in _createOptions)
-            {
-                command += " " + co;
-            }
+            foreach (var co in _createOptions) command += " " + co;
 
             return command;
         }
 
-        public String GetFullCommand()
+        public string GetFullCommand()
         {
             return "create " + GetCommand();
         }
 
-        public IEnumerable<String> GetEntryCommands()
+        public IEnumerable<string> GetEntryCommands()
         {
-            List<String> ret = new List<string>();
-            foreach (var entry in Entries)
-            {
-                ret.Add("add " + _name + " " + entry.GetKeyCommand());
-            }
+            var ret = new List<string>();
+            foreach (var entry in Entries) ret.Add("add " + _name + " " + entry.GetKeyCommand());
 
             return ret;
         }
 
         #endregion
 
-        public static IpSetSet Parse(String[] arguments, IpTablesSystem system, int startOffset = 0)
+        public static IpSetSet Parse(string[] arguments, IpTablesSystem system, int startOffset = 0)
         {
-            IpSetSet set = new IpSetSet(system);
+            var set = new IpSetSet(system);
             var parser = new IpSetSetParser(arguments, set);
 
-            for (int i = startOffset; i < arguments.Length; i++)
-            {
-                i += parser.FeedToSkip(i, startOffset == i);
-            }
+            for (var i = startOffset; i < arguments.Length; i++) i += parser.FeedToSkip(i, startOffset == i);
 
             return set;
         }
 
-        public static IpSetSet Parse(String rule, IpTablesSystem system, int startOffset = 0)
+        public static IpSetSet Parse(string rule, IpTablesSystem system, int startOffset = 0)
         {
-            string[] arguments = ArgumentHelper.SplitArguments(rule);
+            var arguments = ArgumentHelper.SplitArguments(rule);
             return Parse(arguments, system, startOffset);
         }
 
@@ -234,18 +207,12 @@ namespace IPTables.Net.Iptables.IpSet
             if (!(set.MaxElem == MaxElem && set.Name == Name && set.Timeout == Timeout &&
                   set.Type == Type && set.BitmapRange.Equals(BitmapRange) && set.CreateOptions.OrderBy(a => a)
                       .SequenceEqual(CreateOptions.OrderBy(a => a))))
-            {
                 return false;
-            }
 
-            if (size)
-            {
-                return set.HashSize == HashSize;
-            }
+            if (size) return set.HashSize == HashSize;
 
             return true;
         }
-
 
 
         protected void SyncEntriesIp(IEnumerable<IpSetEntry> cidrs)
@@ -263,16 +230,10 @@ namespace IPTables.Net.Iptables.IpSet
                     if (found == BigInteger.Zero)
                     {
                         foreach (var s2 in Entries)
-                        {
                             if (f.Cidr.Contains(s2.Cidr))
-                            {
                                 // size of cidr has changed
                                 if (entriesClone.Remove(s))
-                                {
                                     _system.SetAdapter.DeleteEntry(s2);
-                                }
-                            }
-                        }
 
                         targetEntries[f] = -1;
                     }
@@ -284,10 +245,7 @@ namespace IPTables.Net.Iptables.IpSet
                 }
                 else
                 {
-                    if (entriesClone.Remove(s))
-                    {
-                        _system.SetAdapter.DeleteEntry(s);
-                    }
+                    if (entriesClone.Remove(s)) _system.SetAdapter.DeleteEntry(s);
                 }
             }
 
@@ -295,16 +253,10 @@ namespace IPTables.Net.Iptables.IpSet
             foreach (var s in targetEntries.Where(a => a.Value != 0))
             {
                 if (s.Value > BigInteger.Zero)
-                {
                     foreach (var s2 in entriesClone)
-                    {
                         if (s.Key.Cidr.Contains(s2.Cidr) && s.Key.KeyEquals(s2, false))
-                        {
                             // size of cidr has changed
                             _system.SetAdapter.DeleteEntry(s2);
-                        }
-                    }
-                }
 
                 _system.SetAdapter.AddEntry(s.Key);
             }
@@ -316,43 +268,30 @@ namespace IPTables.Net.Iptables.IpSet
 
             // Go through the system set updating targetEntries if we find something, removing from system if we don't
             foreach (var s in Entries)
-            {
                 if (!targetEntries.Remove(s))
-                {
                     _system.SetAdapter.DeleteEntry(s);
-                }
-            }
 
             // Everything that remains needs to be added
-            foreach (var s in targetEntries)
-            {
-                _system.SetAdapter.AddEntry(s);
-            }
+            foreach (var s in targetEntries) _system.SetAdapter.AddEntry(s);
         }
-        
+
 
         public void SyncEntries(IEnumerable<IpSetEntry> entries)
         {
             try
             {
                 if ((Type & (IpSetType.Ip | IpSetType.Ip2 | IpSetType.Net)) == 0)
-                {
                     // no opportunity for cidr magic
                     SyncEntriesPlain(entries);
-                }
                 else if ((Type & IpSetType.Net) == IpSetType.Net)
-                {
                     SyncEntriesPlain(entries);
-                }
                 else
-                {
                     SyncEntriesIp(entries);
-                }
             }
             catch (Exception ex)
             {
                 throw new IpTablesNetException(
-                    String.Format("An exception occured while adding or removing on entries of set {0} message:{1}",
+                    string.Format("An exception occured while adding or removing on entries of set {0} message:{1}",
                         Name,
                         ex.Message), ex);
             }
