+++ IPTables.Net/Netfilter/TableSync/DefaultNetfilterSync.cs
@@ -6,23 +6,19 @@ using IPTables.Net.Iptables.Adapter.Client;
 
 namespace IPTables.Net.Netfilter.TableSync
 {
-    public class DefaultNetfilterSync: INetfilterSync
+    public class DefaultNetfilterSync : INetfilterSync
     {
         private Func<IpTablesRule, bool> _shouldDelete = null;
 
         public Func<IpTablesRule, bool> ShouldDelete
         {
-            get { return _shouldDelete; }
+            get => _shouldDelete;
             set
             {
                 if (value == null)
-                {
                     _shouldDelete = a => true;
-                }
                 else
-                {
                     _shouldDelete = value;
-                }
             }
         }
 
@@ -30,67 +26,57 @@ namespace IPTables.Net.Netfilter.TableSync
         private IEqualityComparer<IpTablesRule> _comparer = null;
 
 
-        public IEnumerable<String> TableOrder
-        {
-            get;
-            set;
-        } = new List<string> { "raw", "nat", "mangle", "filter" };
+        public IEnumerable<string> TableOrder { get; set; } = new List<string> {"raw", "nat", "mangle", "filter"};
 
         public Func<IpTablesRule, IpTablesRule, bool> RuleComparerForUpdate
         {
-            get { return _ruleComparerForUpdate; }
+            get => _ruleComparerForUpdate;
             set
             {
                 if (value == null)
-                {
                     _ruleComparerForUpdate = (a, b) => false;
-                }
                 else
-                {
                     _ruleComparerForUpdate = value;
-                }
             }
         }
 
-        public DefaultNetfilterSync(Func<IpTablesRule, IpTablesRule, bool> ruleComparerForUpdate = null, Func<IpTablesRule, bool> shouldDelete = null, IEqualityComparer<IpTablesRule> comparer = null)
+        public DefaultNetfilterSync(Func<IpTablesRule, IpTablesRule, bool> ruleComparerForUpdate = null,
+            Func<IpTablesRule, bool> shouldDelete = null, IEqualityComparer<IpTablesRule> comparer = null)
         {
             ShouldDelete = shouldDelete;
             RuleComparerForUpdate = ruleComparerForUpdate;
             _comparer = comparer ?? new IpTablesRule.ValueComparison();
-        } 
+        }
 
         public void SyncChainRules(IIPTablesAdapterClient client, IEnumerable<IpTablesRule> with, IpTablesChain chain)
         {
             //Copy the rules
             var currentRules = new List<IpTablesRule>(chain.Rules);
 
-            
+
             int i = 0, len = with.Count();
 
-            bool shouldUpdate = currentRules.Count == len;
-            foreach (IpTablesRule cR in currentRules)
+            var shouldUpdate = currentRules.Count == len;
+            foreach (var cR in currentRules)
             {
                 //Delete any extra rules
                 if (i == len)
                 {
-                    if (_shouldDelete(cR))
-                    {
-                        cR.DeleteRule(client);
-                    }
+                    if (_shouldDelete(cR)) cR.DeleteRule(client);
                     continue;
                 }
 
                 //Get the rule for comparison
-                IpTablesRule withRule = with.ElementAt(i);
+                var withRule = with.ElementAt(i);
 
-                bool eq = _comparer.Equals(cR,withRule);
+                var eq = _comparer.Equals(cR, withRule);
                 if (eq)
                 {
                     //No need to make any changes
                     i++;
                     continue;
                 }
-                
+
                 //Debug:
                 if (_ruleComparerForUpdate(cR, withRule) || shouldUpdate)
                 {
@@ -101,15 +87,12 @@ namespace IPTables.Net.Netfilter.TableSync
                 else
                 {
                     // Don't delete if this is non deletable
-                    if (_shouldDelete(cR))
-                    {
-                        cR.DeleteRule(client);
-                    }
+                    if (_shouldDelete(cR)) cR.DeleteRule(client);
                 }
             }
 
             //Get rules to be added
-            foreach (IpTablesRule rR in with.Skip(i))
+            foreach (var rR in with.Skip(i))
             {
                 var newRule = rR.ShallowClone();
                 newRule.Chain = chain;
@@ -119,7 +102,6 @@ namespace IPTables.Net.Netfilter.TableSync
 
         public void SyncChains()
         {
-            
         }
     }
 }
